refactor on Entity class to move to a model namespace
