+++ lib/shorty.rb
@@ -20,4 +20,4 @@ require "./lib/shorty/controllers/show.rb"
 require "./lib/shorty/application.rb"
 require "./lib/shorty/shortcode/validator.rb"
 require "./lib/shorty/shortcode/generator.rb"
-require "./lib/shorty/shorty_entity.rb"
+require "./lib/shorty/models/shorty.rb"
